requirements.txt
python-dotenv
mysql.connector
mysql-connector-python
requests
json
dotenv